Complete rewrite of the AccountReputation script
Make sure you also run the latest .SQL file as well.

This was a complete rewrite of the Accountwide reputation logic to convert the various reputation base values for each Race found in the Faction.dbc file.  Also kept the script true to keeping the reputation within the respective faction.  Horde factions are only shared to horde characters.  Alliance factions are only shared to alliance characters.  Neutral factions are shared to both.