Compatibility and caching improvements for Playerbots
When I wrote this systems initially I was using Trickerer's NPCBots.  After recently migrating over to Playerbots, i've learned that the bots use actual player accounts and with hundreds or thousands of those playerbot accounts all logging in simultaneously or triggering player data, it would put a ton of stress on these systems which I wasn't aware of since I wasn't using Playerbots.  People have commented about massive slowdowns when using these systems that I would constantly struggle to replicate and I think it has to do with the Playerbot/NPCBot difference since NPCBots don't use player data or player accounts.

Now that I am aware of this, this is the first big step to making these systems way more compatible/performant friendly when using a Playerbots fork.  The idea is that I created a new helper script `00_AccountWideUtils.lua` with some helper functions that are called from each of the script before executing their main accountwide syncing logic.  It will check to see if it is a Playerbot account or an actual player account before executing.  It will exclude and skip any playerbots accounts and should continue to behave as normal for actual players.  This means that the `00_AccountWideUtils.lua` file is REQUIRED and must be used alongside any of the other accountwide files due to the use of helper functions.

I will continue to monitor and test this on a Playerbots fork for any issues and will continue to make improvements if needed.  Please feel free to contact me on discord @aldori or open up an Issue/Bug on this Github repo is you see any issues.