Reputation - Use proper positive and negative clamping values for adjusted standing
This fixes a bug where the previous clamping approach of (-42000, 42000) had a ceiling of 42000.  While at first glance this would mean (Hated, Exalted), I wasn't taking into account the base reputation values.

For example, Brood of Nozdormu (faction 910) has a base value of -42000 for all races, meaning that everyone starts out as Hated.  In order to reach Exalted (42000), you need to earn +84000 reputation.  Since the previous ceiling was capped at 42000, this would mean that your Brood of Nozdormu reputation was always revert back to Neutral (0) if you had previously obtained Exalted.  With the ceiling clamped at 84000, this allows for that specific scenario of starting with the lowest possible base value of -42000 to reach exalted 42000.

Same logic applies for the negative clamping value changing that from -42000 to -46000.  Since the highest positive base value is 4000, assuming you go all the way down to Hated (-42000) from that highest base value, you would have to lose -46000 reputation to do so.  With the previous clamp of -42000, this would always revert you to Hostile (-38000) due to the floor clamp.  Although this specific scenario is even less likely, it's just added as a just in case.